multi_agent_codegen/requirements.txt
# Core dependencies
python>=3.8
pyyaml>=6.0
pytest>=7.4.0

# Web Framework
flask>=2.3.0
werkzeug>=2.3.0

# LLM Integration (reused from parent project)
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0

# LangGraph for workflow orchestration
langgraph>=0.0.20
langchain>=0.1.0

# Optional: Docker support for sandbox
# docker>=6.0.0  # Uncomment if using Docker sandbox

# Development tools
black>=23.0.0
flake8>=6.0.0

